backend_project/requirements.txt
alembic==1.13.1
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
greenlet==3.0.3
inflection==0.5.1
iniconfig==2.0.0
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-django==4.8.0
pytest-dotenv==0.5.2
pytest-sqlalchemy==0.2.1
python-dotenv==1.0.1
SQLAlchemy==2.0.30
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
typing_extensions==4.12.2
